+++ runtime/moonbase/src/governance/councils.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with Moonbeam.  If not, see <http://www.gnu.org/licenses/>.
 
-//! Councils from old governance config
+//! Councils for Gov1
 
 use super::*;
