+++ runtime/moonbase/src/governance/democracy.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with Moonbeam.  If not, see <http://www.gnu.org/licenses/>.
 
-//! Old democracy config
+//! Democracy config for Gov1
 
 use super::councils::*;
 use crate::*;
